Refactored persistance exception
