+++ scb-persistenceproviders/elasticsearch-persistenceprovider/src/main/java/io/securecodebox/persistence/elasticsearch/ElasticSearchPersistenceProvider.java
@@ -264,12 +264,6 @@ public class ElasticSearchPersistenceProvider implements PersistenceProvider {
         }
     }
 
-    private static class InvalidContextNameForElkIndex extends PersistenceException{
-        public InvalidContextNameForElkIndex(String contextName) {
-            this.message = "Cannot create custom elasticsearch index for context name '" + contextName + "' as it contains reserved characters. Please choose a different context name.";
-        }
-    }
-
     private String transformContextForElasticsearchIndexCompatability() {
         if (context != null) {
             String contextIndex = context.toLowerCase().replace(" ", "_") + "_";
@@ -279,7 +273,7 @@ public class ElasticSearchPersistenceProvider implements PersistenceProvider {
                 return contextIndex;
             } catch (InvalidIndexNameException e) {
                 LOG.error("Context name contains chars which are invalid to be a elasticsearch index name. Please change the context name so that a context specific index can be created.");
-                throw new InvalidContextNameForElkIndex(context);
+                throw new InvalidContextNameForElkIndex("Cannot create custom elasticsearch index for context name '" + context + "' as it contains reserved characters. Please choose a different context name.");
             }
         }
